Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -2,7 +2,7 @@
 int main() {
     int a,b,c;
     scanf("%d %d %d",&a,&b,&c);
-    int sum = a+b+c;
+    float sum = (float)a+b+c;
     printf("Average: %.2f",sum/3);
     return 0;
 }
